subscriber/requirements.txt
paho-mqtt
redis
pymongo
pyyaml
pyaml-env
